Update README.md
This project describes the  sales and reoprt of the buyers and product in the diwali season. In this project I have learned the data cleaning , data visualization and exploration. 
Tools ie python (pandas, matplotlib, seabord), google colab notebook and excel for data analysis and visualization . 
The excel data sheet of for the diwali sales data is uploaded.